Struct anpp::Packet

pub struct Packet { /* fields omitted */ }

An arbitrary chunk of bytes with a user-defined ID.

Implementations

impl Packet

pub const MAX_PACKET_SIZE: usize

The maximum size a packet can be.

pub fn new(id: u8) -> Packet

Create a new empty Packet.

pub fn with_data(id: u8, content: &[u8]) -> Result<Packet, InsufficientCapacity>

Create a Packet already filled with data.

pub fn id(&self) -> u8

Get the Packet's identifier.

pub fn contents(&self) -> &[u8]

Get the contents of the Packet.

pub fn len(&self) -> usize

How long is the Packet's contents?

pub fn is_empty(&self) -> bool

Is this Packet empty?

pub fn remaining_capacity(&self) -> usize

How many more bytes can be added to this Packet before it is full?

pub fn total_length(&self) -> usize

The number of bytes this Packet will take up when encoded and written to a buffer.

pub fn write_to_buffer(&self, buffer: &mut [u8]) -> Result<usize, EncodeError>

Encode a Packet and copy it into the provided buffer.

Note

The minimum required buffer size can be determined via the total_length() method. If the buffer isn't big enough, nothing will be written to the buffer and this method will fail with an InsufficientCapacity error.

pub fn push_data(&mut self, data: &[u8]) -> Result<(), InsufficientCapacity>

Try to append some data to the Packet's contents, bailing without writing anything if there isn't enough space.